Fixed issue by changing all size_t to int. Made print_vector function.

+++ data_manager.h
@@ -5,27 +5,27 @@
 #include <iostream>
 #include <algorithm>
 
-std::vector<size_t> generate_dataset(size_t dataset_size, size_t start_value = 0)
+std::vector<int> generate_dataset(int dataset_size, int start_value = 0)
 {
-    std::vector<size_t> dataset(dataset_size);
+    std::vector<int> dataset(dataset_size);
 
     #pragma omp parallel for
-    for (size_t i = 0; i < dataset_size; i++) {
+    for (int i = 0; i < dataset_size; i++) {
         dataset[i] = i + start_value;
     }
 
     return dataset;
 }
 
-std::vector<size_t> generate_random_dataset(size_t dataset_size)
+std::vector<int> generate_random_dataset(int dataset_size)
 {
-    std::vector<size_t> dataset(dataset_size);
+    std::vector<int> dataset(dataset_size);
 
     #pragma omp parallel for
-    for (size_t i = 0; i < dataset_size; i++) {
+    for (int i = 0; i < dataset_size; i++) {
         std::random_device rd;
         std::mt19937 gen(rd());
-        std::uniform_int_distribution<size_t> dis(0, dataset_size);
+        std::uniform_int_distribution<int> dis(0, dataset_size);
 
         dataset[i] = dis(gen);
     }
@@ -34,7 +34,7 @@ std::vector<size_t> generate_random_dataset(size_t dataset_size)
 }
 
 // Fisher-Yates shuffle -> shuffles the given dataset
-void shuffle_dataset(std::vector<size_t>& dataset)
+void shuffle_dataset(std::vector<int>& dataset)
 {
     std::random_device rd;
     std::mt19937 gen(rd());
@@ -42,7 +42,7 @@ void shuffle_dataset(std::vector<size_t>& dataset)
     std::shuffle(dataset.begin(), dataset.end(), gen);
 }
 
-void save_dataset(const std::vector<size_t>& dataset, std::string filename)
+void save_dataset(const std::vector<int>& dataset, std::string filename)
 {
     std::ofstream file(filename, std::ios::binary);
 
@@ -51,7 +51,7 @@ void save_dataset(const std::vector<size_t>& dataset, std::string filename)
         return;
     }
 
-    file.write(reinterpret_cast<const char*>(dataset.data()), dataset.size() * sizeof(size_t));
+    file.write(reinterpret_cast<const char*>(dataset.data()), dataset.size() * sizeof(int));
     file.close();
 }
 
@@ -66,11 +66,11 @@ std::vector<int> load_dataset(const std::string& filename)
 
     // Verkrijg de grootte van het bestand om te bepalen hoeveel data we moeten inleiden
     file.seekg(0, std::ios::end);  // Ga naar het einde van het bestand
-    size_t file_size = file.tellg();  // Verkrijg bestandsgrootte
+    int file_size = file.tellg();  // Verkrijg bestandsgrootte
     file.seekg(0, std::ios::beg);  // Ga terug naar het begin van het bestand
 
     // Bepaal hoeveel integers er in het bestand staan
-    size_t num_elements = file_size / sizeof(size_t);
+    int num_elements = file_size / sizeof(int);
 
     std::vector<int> dataset(num_elements);  // Creëer een vector voor de data
